+++ src/room.py
@@ -1,5 +1,3 @@
-import math
-
 from seika.math import Vector2
 from seika.node import Node2D, CollisionShape2D
 
@@ -72,39 +70,3 @@ class Room:
 
     def __repr__(self):
         return f"(position = {self.position}, size = {self.size})"
-
-
-class RoomManager:
-    _instance = None
-
-    def __new__(cls, *args, **kwargs):
-        if not cls._instance:
-            cls._instance = object.__new__(cls)
-            cls.current_room = None
-            cls.rooms = {}
-            cls.wall_colliders = None
-            cls.room_doors = None
-        return cls._instance
-
-    def add_room(self, room: Room) -> None:
-        self.rooms[f"{room.position.x}-{room.position.y}"] = room
-
-    def set_current_room(self, position: Vector2) -> None:
-        pos_key = f"{position.x}-{position.y}"
-        if pos_key in self.rooms:
-            self.current_room = self.rooms[pos_key]
-
-    def get_room(self, position: Vector2) -> Room:
-        return self.rooms[f"{position.x}-{position.y}"]
-
-    def get_grid_position(self, position: Vector2) -> Vector2:
-        return Vector2(
-            math.floor(position.x / ProjectProperties.BASE_RESOLUTION.x),
-            math.floor(position.y / ProjectProperties.BASE_RESOLUTION.y),
-        )
-
-    def get_world_position(self, grid_position: Vector2) -> Vector2:
-        return Vector2(
-            math.floor(grid_position.x * ProjectProperties.BASE_RESOLUTION.x),
-            math.floor(grid_position.y * ProjectProperties.BASE_RESOLUTION.y),
-        )
